refactor(server): extract Zernio client into src/server/zernio.js
Pure move, no behavior change. Relocates the Zernio social-publishing cluster:
callZernio (the single HTTP primitive), zernioPublish, getOrCreateZernioProfile,
ensureZernioProfile, and the dev-only zernioGuard.

zernio.js imports pool (db.js) and verifyBrandAccess (auth.js) — clean layering
on the already-extracted modules, no cycle (server.js -> zernio.js -> auth.js ->
db.js). server.js re-imports all five (callZernio alone has 24 call sites).

no-undef gate clean (every call site + verifyBrandAccess/pool resolve across
both files), route inventory unchanged (213). Adds test/zernio.test.js covering
zernioGuard's prod-host / admin-password / env-var gates. Full suite: 37 green.
